requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
anthropic==0.39.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
Pillow==10.4.0
rembg==2.0.59
